[#37] Feat: 유저모듈로 마이페이지 유저통신 파일이동
