+++ Projects/Data/User/Sources/DTO/UserInfoDTO+.swift
@@ -6,14 +6,14 @@
 //
 
 import Foundation
-import MyPageDomain
+import UserDomain
 
 extension UserInfoDTO {
     var toDomain: UserInfo {
         UserInfo(
             userId: userId,
             nickname: nickname,
-            profileImageCode: profileImageCode,
+            profileType: ProfileType(rawValue: profileImageCode) ?? .purple,
             isReservationAlarmSetting: reservationAlarmSetting,
             isKokAlarmSetting: kokAlarmSetting,
             totalReservations: statistics.totalReservations,
