+++ Projects/Domain/User/Sources/Repository/UserRepository.swift
@@ -8,6 +8,8 @@
 import Foundation
 
 public protocol UserRepository {
+    typealias StatusCode = Int
+    
     func signUp(
         deviceId: String,
         nickname: String,
@@ -19,4 +21,11 @@ public protocol UserRepository {
     ) async throws
     
     func getProfileList() async throws -> [ProfileEntity]
+    func requestUserInfo() async throws -> UserInfo
+    func requestUpdateUserInfo(
+        nickname: String?,
+        profileImageCode: String?,
+        isReservationAlarm: Bool?,
+        isKokAlarm: Bool?
+    ) async throws -> StatusCode
 }
