+++ Projects/Domain/User/Sources/UseCase/UserUseCaseImpl.swift
@@ -1,0 +1,82 @@
+//
+//  UserUseCaseImpl.swift
+//  UserDomain
+//
+//  Created by Enes on 7/3/25.
+//
+
+import Foundation
+import HGCommon
+import HGLogger
+
+public final class UserUseCaseImpl: UserUseCase {
+    
+    private let userRepo: UserRepository
+    private let keychain: KeychainManagerable
+    
+    public init(
+        userRepo: UserRepository,
+        keychain: KeychainManagerable
+    ) {
+        self.userRepo = userRepo
+        self.keychain = keychain
+    }
+    
+    public func signUp(
+        deviceId: String,
+        nickname: String,
+        profileType: ProfileType
+    ) async throws {
+        
+        /// 회원가입
+        let response = try await userRepo.signUp(
+            deviceId: deviceId,
+            nickname: nickname,
+            profileType: profileType
+        )
+        LoggerUtil.log("회원가입 성공 UserId: \(response.userId) ")
+        
+        /// JWT 저장
+        try await keychain.addKeychain(key: .accessToken, value: response.accessToken)
+        
+        /// FCM 등록
+        let fcmToken = try await keychain.readKeychain(key: .fcmToken)
+        try await userRepo.updateFCM(fcmToken: fcmToken)
+        LoggerUtil.log("FCM 등록 성공 FCM Token: \(fcmToken) ")
+    }
+    
+    public func getProfileList() async throws -> [ProfileEntity] {
+        try await userRepo.getProfileList()
+    }
+    
+    public func validateNickname(nickname: String) -> Bool {
+        return !nickname.contains { c in
+            c.isEmoji || c.isWhitespace || c.isNewline
+        }
+    }
+    
+    public func requestUserInfo() async throws -> UserInfo {
+        try await userRepo.requestUserInfo()
+    }
+    
+    public func requestUpdateUserInfo(
+        nickname: String?,
+        profileImageCode: String?,
+        isReservationAlarm: Bool?,
+        isKokAlarm: Bool?
+    ) async throws -> Bool {
+        let statusCode = try await userRepo.requestUpdateUserInfo(
+            nickname: nickname,
+            profileImageCode: profileImageCode,
+            isReservationAlarm: isReservationAlarm,
+            isKokAlarm: isKokAlarm
+        )
+        return statusCode == 200
+    }
+}
+
+fileprivate extension Character {
+    var isEmoji: Bool {
+        unicodeScalars.contains { $0.properties.isEmoji }
+    }
+}
